Fixes the citation bugs
The build broke because of unknown unicode characters. Due to this the
compilation stopped.

The unknown characters \u202F (Thin Spaces) were removed. Furthermore
the `°` in the listing was added explicitly.

In addition the Makefile was manipulated to make sure LaTeX runs
multiple times each call, to ensure, that references get linked
correctely.